+++ contracts/modules/VersionModule.sol
@@ -11,7 +11,7 @@ abstract contract VersionModule is IERC3643Base {
      * @dev
      * Get the current version of the smart contract
      */
-    string private constant VERSION = "0.2.0";
+    string private constant VERSION = "0.3.0";
 
     /* ============ Events ============ */
     /*//////////////////////////////////////////////////////////////